Refactor board-related classes: clean up code by removing unused imports and methods, improve formatting, and enhance documentation for clarity.
