+++ src/main/java/com/gruppe24/observerpattern/EventType.java
@@ -1,5 +1,9 @@
 package com.gruppe24.observerpattern;
 
+/**
+ * Enum representing different types of events in the game.
+ * This enum is used to categorize events that can occur during the game.
+ */
 public enum EventType {
   DICE_ROLLED,
   PLAYER_MOVED,
