+++ src/main/java/com/gruppe24/observerpattern/GameSubject.java
@@ -3,11 +3,18 @@ package com.gruppe24.observerpattern;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * The GameSubject class implements the GameObservable interface and is responsible for managing
+ * the list of observers and notifying them of events.
+ */
 public class GameSubject implements GameObservable {
 
   private final List<GameObserver> observers;
 
-
+  /**
+   * Constructor for the GameSubject class.
+   * Initializes the list of observers.
+   */
   public GameSubject() {
     observers = new ArrayList<>();
   }
@@ -23,12 +30,10 @@ public class GameSubject implements GameObservable {
     }
   }
 
-
   @Override
   public void removeListener(GameObserver observer) {
     observers.remove(observer);
   }
-
 
   @Override
   public void notifyListener(EventType event, Object... args) {
