+++ src/main/java/com/gruppe24/utils/StyleUtils.java
@@ -90,70 +90,4 @@ public class StyleUtils {
     }
     node.setStyle(panelStyle());
   }
-
-  //--------Colour---------
-
-  /**
-   * Method to style the text in red.
-   *
-   * @return the ANSI escape code for red text
-   */
-  public static String textRed() {
-    return "\u001B[31m";
-  }
-
-  /**
-   * Method to style the text in blue.
-   *
-   * @return the ANSI escape code for blue text
-   */
-  public static String textBlue() {
-    return "\u001B[34m";
-  }
-
-  /**
-   * Method to style the text in green.
-   *
-   * @return the ANSI escape code for green text
-   */
-  public static String textGreen() {
-    return "\u001B[32m";
-  }
-
-  /**
-   * Method for resetting the text color.
-   *
-   * @return the ANSI escape code for resetting text color
-   */
-  public static String textReset() {
-    return "\u001B[0m";
-  }
-
-  /**
-   * Method for printing a rose in ASCII art.
-   *
-   * @return the ASCII art of a rose
-   */
-  public static String textRose() {
-    return textRed() + """
-        
-            _,--._.-,
-           /\\_r-,\\_ )
-        .-.) _;='_/ (.;
-         \\ \\'     \\/S )
-          L.'-. _.'|-'
-         <_`-'\\'_.'/
-           `'-._( \\
-        """
-        + textGreen()
-        + """
-                ___   \\\\,      ___
-                \\ .'-. \\\\   .-'_. /
-                 '._' '.\\\\/.-'_.'
-                    '--``\\('--'
-                          \\\\
-                          `\\\\,
-                            \\|⠀⠀⠀⠀⠀⠀⠀
-            """ + textReset();
-  }
 }
